Add files via upload
This chatbot project is a web-based application designed to provide users with an interactive and engaging experience. Utilizing Flask as the backend framework, this chatbot can respond to user queries in real-time, simulating a conversation. The frontend is built with HTML, CSS (Bootstrap), and JavaScript, creating a responsive and user-friendly interface.

Features:

    Real-Time Chat: Users can send messages and receive responses instantly.
    Responsive Design: The application is optimized for both desktop and mobile devices.
    Customizable: Users can easily change the chatbot's appearance, including icons and background colors.
    AJAX Integration: Utilizes AJAX for seamless communication between the frontend and backend without page reloads.

Technologies Used:

    Frontend: HTML, CSS (Bootstrap), JavaScript, jQuery
    Backend: Python, Flask
    Deployment: Can be easily deployed on any web server that supports Python.